+++ src/app/app.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { NavigationEnd, Router } from '@angular/router';
+import { ActivatedRoute, NavigationEnd, Router } from '@angular/router';
 import { Store } from '@ngrx/store';
 
 import { HandleSubscription } from 'common/handle-subscription';
@@ -28,6 +28,7 @@ export class AppComponent extends HandleSubscription implements OnInit {
 
   constructor(
     private store: Store<AppState>,
+    private route: ActivatedRoute,
     private router: Router,
     private auth: AuthService,
     private session: SessionService
@@ -44,6 +45,7 @@ export class AppComponent extends HandleSubscription implements OnInit {
       .select('state', 'common', 'info')
       .subscribe((info: Info) => {
         if (
+          !this.isOauth() &&
           environment.adControllerUrl &&
           this.MODE_INITIALIZATION === info?.mode
         ) {
@@ -75,5 +77,9 @@ export class AppComponent extends HandleSubscription implements OnInit {
 
   loadInfo(): void {
     this.store.dispatch(new LoadInfo());
+  }
+
+  isOauth(): boolean {
+    return undefined !== this.route.snapshot.queryParams.redirect_uri;
   }
 }