+++ src/app/auth/auth.component.ts
@@ -41,7 +41,6 @@ export class AuthComponent extends HandleSubscription {
   }
 
   ngOnInit(): void {
-    console.log(this.source);
     this.store
       .select('state', 'common', 'info')
       .pipe(take(1))