+++ src/app/store/admin/admin.actions.ts
@@ -4,7 +4,6 @@ import {
   AdvertiserInfo,
   License,
   PublisherInfo,
-  UserBanDetails,
   UserInfo,
 } from 'models/settings.model';
 
@@ -27,15 +26,6 @@ export const LOAD_ADMIN_SETTINGS_FAILURE = 'Admin settings loading failure';
 export const GET_LICENSE = 'Get license';
 export const GET_LICENSE_SUCCESS = 'Get license success';
 export const GET_LICENSE_FAILURE = 'Get license failure';
-
-export const BAN_USER = 'Ban user';
-export const BAN_USER_SUCCESS = 'Ban user success';
-
-export const UNBAN_USER = 'Unban user';
-export const UNBAN_USER_SUCCESS = 'Unban user success';
-
-export const DELETE_USER = 'Delete user';
-export const DELETE_USER_SUCCESS = 'Delete user success';
 
 export class LoadUsers implements Action {
   readonly type: string = LOAD_USERS;
@@ -149,42 +139,6 @@ export class GetLicenseFailure implements Action {
   constructor(public payload?: string) {}
 }
 
-export class BanUser implements Action {
-  readonly type: string = BAN_USER;
-
-  constructor(public payload: UserBanDetails) {}
-}
-
-export class BanUserSuccess implements Action {
-  readonly type: string = BAN_USER_SUCCESS;
-
-  constructor(public payload?: UserInfo) {}
-}
-
-export class UnbanUser implements Action {
-  readonly type: string = UNBAN_USER;
-
-  constructor(public payload: number) {}
-}
-
-export class UnbanUserSuccess implements Action {
-  readonly type: string = UNBAN_USER_SUCCESS;
-
-  constructor(public payload?: UserInfo) {}
-}
-
-export class DeleteUser implements Action {
-  readonly type: string = DELETE_USER;
-
-  constructor(public payload: number) {}
-}
-
-export class DeleteUserSuccess implements Action {
-  readonly type: string = DELETE_USER_SUCCESS;
-
-  constructor(public payload?: number) {}
-}
-
 export type actions =
   | LoadUsers
   | LoadUsersSuccess
@@ -200,10 +154,4 @@ export type actions =
   | LoadAdminSettingsFailure
   | GetLicense
   | GetLicenseSuccess
-  | GetLicenseFailure
-  | BanUser
-  | BanUserSuccess
-  | UnbanUser
-  | UnbanUserSuccess
-  | DeleteUser
-  | DeleteUserSuccess;
+  | GetLicenseFailure;